POTMS/requirements.txt
anyio==4.12.0
asgiref==3.9.1
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
firebase_admin==7.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.1.4
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
whitenoise==6.11.0
qrcode[pil]>=7.4.2

# Security packages
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0
psycopg2-binary>=2.9.9
